Add files via upload
I add a new .root file with the events that I want to analyze, in this case it is a mixture of Higgs events and ttbar events. The file has been cutted from a bigger one in order to have just the information that I want to use for the analysis and to have a file of a size that I can upload on the repository.